requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.9.1
Werkzeug==3.0.1
python-dotenv==1.0.1
email-validator==2.1.0.post1
markdown==3.5.2
pytest==8.0.0
pytest-flask==1.3.0
coverage==7.4.1
black==24.1.1
flake8==7.0.0 